requirements.txt
absl-py==2.1.0
appdirs==1.4.4
asttokens==2.4.1
astunparse==1.6.3
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.1
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
docker-pycreds==0.4.0
executing==2.0.1
filelock==3.13.1
flatbuffers==23.5.26
fonttools==4.47.2
fsspec==2024.2.0
gast==0.5.4
geomloss==0.2.6
gitdb==4.0.11
GitPython==3.1.41
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
grpcio==1.60.1
h5py==3.10.0
idna==3.6
ipykernel==6.29.1
ipython==8.21.0
ipywidgets==8.1.2
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyterlab_widgets==3.0.10
keopscore==2.2.2
keras==2.15.0
kiwisolver==1.4.5
lazy_loader==0.4
libclang==16.0.6
Markdown==3.5.2
MarkupSafe==2.1.5
matplotlib==3.8.2
matplotlib-inline==0.1.6
ml-dtypes==0.2.0
mne==1.7.0
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.3
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.2
pandas==2.2.0
parso==0.8.3
pillow==10.2.0
platformdirs==4.2.0
pooch==1.8.1
prompt-toolkit==3.0.43
protobuf==4.23.4
psutil==5.9.8
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pybind11==2.11.1
Pygments==2.17.2
pykeops==2.2.2
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2024.1
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.2
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.4.0
scipy==1.12.0
seaborn==0.13.2
sentry-sdk==1.40.0
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
stack-data==0.6.3
sympy==1.12
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
threadpoolctl==3.2.0
torch==2.2.0
torchaudio==2.2.0
torchsummary==1.5.1
torchvision==0.17.0
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.2.0
wandb==0.16.3
wcwidth==0.2.13
Werkzeug==3.0.1
widgetsnbextension==4.0.10
wrapt==1.14.1
